# docs/server/openapi.v1.yaml
openapi: 3.1.0
info:
  title: SpeechKit Server API
  version: 1.0.0
  description: |
    Stable v1 HTTP and WebSocket contract for the SpeechKit Server.
    The canonical server API prefix is `/v1/`; `/api/v1/` remains
    available as a compatibility alias. These endpoints are covered by the same
    backwards-compatibility guarantees as the Framework's public
    `pkg/speechkit` packages — fields are additive across minor releases; no
    rename or removal without a version bump.

    Ships with the container image
    `ghcr.io/kombifyio/speechkit-server:v0.26.0` and later. Earlier images
    expose the same surface but were Server-internal pre-v0.26.

  contact:
    url: https://github.com/kombifyio/SpeechKit
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.speechkit.dev
    description: Replace with your deployment's URL.

# ── Auth ────────────────────────────────────────────────────────────────────
# Default server deployments require bearer auth. `auth_mode = "none"` is only
# intended for loopback/dev deployments and must not be used on public binds.
security:
  - {}
  - bearer: []
  - edgeHmac: []
paths:
...
  /v1/assist/process:
    post:
      operationId: processAssist
      summary: STT → codeword/utility/LLM → optional TTS
      description: |
        Accepts either an audio payload (transcribed via STT first) or a
        text transcript directly. Runs the Framework's Assist pipeline,
        which routes to a tool intent or the LLM agent flow, and returns
        the result plus optional TTS audio as base64 in the JSON response.

        Tool execution is NOT done server-side; clients receive an
        `action: "execute"` signal plus the matched shortcut name and
        perform the action themselves.
      tags: [assist]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio:
                  type: string
                  format: binary
                text:
                  type: string
                locale:
                  type: string
                selection:
                  type: string
                context:
                  type: string
                tts:
                  type: string
                  description: '"true"/"false" — opt out of TTS audio in the response.'
                tts_format:
                  type: string
                  enum: [mp3, wav, opus]
                tts_voice:
                  type: string
          application/json:
            schema: { $ref: '#/components/schemas/AssistJSONRequest' }
      responses:
        '200':
          description: Pipeline succeeded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssistResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422':
          description: STT returned an empty transcript (audio contained no speech).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ProviderUnavailable' }
  /v1/assist/self-test:
    post:
      operationId: selfTestAssist
      summary: Run a text-only Assist self-test
      description: |
        Runs a fixed text-only Assist prompt with TTS output ignored. Use this
        to separate client/proxy issues from server-side LLM/provider
        configuration failures.
      tags: [assist]
      responses:
        '200':
          description: Assist pipeline produced a non-empty response.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssistSelfTestResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ProviderUnavailable' }

  # ── Voice Agent ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: |
        Static deployment-wide token from `$SPEECHKIT_SERVER_TOKEN`. The
        identity attached to the request context is `service@default/internal`.
        Suitable for internal service-to-service callers, product backends, or
        automation services on a private network.
    edgeHmac:
      type: apiKey
      in: header
      name: X-Edge-Auth-Hmac
      description: |
        HMAC-SHA256 of `userID + "\n" + orgID + "\n" + plan + "\n" + role` using
        `$EDGE_AUTH_SECRET`. Must be sent alongside `X-Edge-User-Id`,
        `X-Edge-Org-Id`, `X-Edge-Plan`, and optional `X-Edge-Role`. Used
        when fronted by a Cloudflare Worker / reverse proxy that has
        already validated a user JWT.
...
  responses:
...
    BadRequest:
      description: The request body could not be decoded or failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Unauthenticated:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    PayloadTooLarge:
      description: Request body exceeds the configured maximum upload size.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UnsupportedMediaType:
      description: Content-Type is not one this endpoint accepts.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Rate limit exceeded. Honour the `Retry-After` header.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ProviderUnavailable:
      description: |
        No upstream provider was able to satisfy the request. Typical for
        STT/Genkit/TTS outages, or when the deployment has no provider
        keys configured for the requested mode.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    AssistJSONRequest:
      type: object
      properties:
        text: { type: string }
        audio_base64: { type: string }
        format:
          type: string
          enum: [wav, mp3, pcm16]
        locale: { type: string }
        selection: { type: string }
        context: { type: string }
        tts: { type: boolean }
        tts_format:
          type: string
          enum: [mp3, wav, opus]
        tts_voice: { type: string }
        tts_speed:
          type: number
          format: double
    AssistResponse:
      type: object
      properties:
        text: { type: string }
        speak_text: { type: string }
        action:
          type: string
          enum: [respond, execute, silent, shortcut]
        locale: { type: string }
        shortcut: { type: string }
        surface: { type: string }
        kind: { type: string }
        transcript: { type: string }
        audio_base64: { type: string }
        audio_format: { type: string }
        latency_ms: { type: integer }
        source:
          $ref: '#/components/schemas/SourceMeta'
    AssistSelfTestResponse:
      type: object
      properties:
        status: { type: string, enum: [ok] }
        text: { type: string }
        action: { type: string }
        locale: { type: string }
        latency_ms: { type: integer }
        details:
          type: object
          additionalProperties: true
...
    SourceMeta:
      type: object
      properties:
        format: { type: string }
        sample_rate: { type: integer }
        channels: { type: integer }
        duration_ms: { type: integer }
...